Adding entry cache to `lazyObj` for storing already type converted entries for subsequent reuse.

Before this change, `.Find()` invocations on `lazyObj` instances where the searched for `ref` contains multiple variables would cause multiple type conversions for the same object entry; which can be very costly for large objects.