@srenatus 2023-06-20
I'd love to see a benchstat comparison for this, main vs PR 😅 But I can do that myself tomorrow. Thanks for adding a benchmark!