@johanfylling 2023-06-20
Some numbers I've collected: 

<img width="1095" alt="cpu" src="https://github.com/open-policy-agent/opa/assets/16031288/16c29d78-6853-4153-af79-7bba2ceb799f">

- blue: `main` with lazy objects turned on
- red: `main` with lazy objects turned off
- yellow: the fix in this PR

x-axis is logarithmic.

Data was collected by running `opa test --bench` on the following policy:

```rego
package example

default allow = false

allow {
    input.x == data.allowed[_][_]
}

test_1 {
	allow with input as {"x": "1"}
}

test_5 {
	allow with input as {"x": "5"}
}

test_10 {
	allow with input as {"x": "10"}
}

test_15 {
	allow with input as {"x": "15"}
}

test_20 {
	allow with input as {"x": "20"}
}

test_25 {
	allow with input as {"x": "25"}
}

test_30 {
	allow with input as {"x": "30"}
}

test_70 {
	allow with input as {"x": "70"}
}

test_100 {
	allow with input as {"x": "100"}
}

test_na {
	not allow with input as {"x": "1000"}
}
```

and data:

```json
{
    "allowed": {
        "a": ["1", "2"],
        "b": ["3"],
        "c": ["4", "5", "6", "7", "8"],
        "d": ["9", "10", "11", "12", "13", "14"],
        "e": ["15", "16", "17"],
        "f": ["18", "19", "20"],
        "g": ["21", "22", "23", "24", "25", "26"],
        "h": ["27", "28", "29", "30"],
        "x": ["1", "2", "3", "4", "5", "6", "7", "8", "9", "10", "11", "12", "13", "14", "15", "16", "17", "18", "19", "20", "21", "22", "23", "24", "25", "26", "27", "28", "29", "30", "31", "32", "33", "34", "35", "36", "37", "38", "39", "40", "41", "42", "43", "44", "45", "46", "47", "48", "49", "50", "51", "52", "53", "54", "55", "56", "57", "58", "59", "60", "61", "62", "63", "64", "65", "66", "67", "68", "69", "70"],
        "y": ["71", "72", "73", "74", "75", "76", "77", "78", "79", "80", "81", "82", "83", "84", "85", "86", "87", "88", "89", "90", "91", "92", "93", "94", "95", "96", "97", "98", "99", "100", "101", "102", "103", "104", "105", "106", "107", "108", "109", "110", "111", "112", "113", "114", "115", "116", "117", "118", "119", "120", "121", "122", "123", "124", "125", "126", "127", "128", "129", "130", "131", "132", "133", "134", "135", "136", "137", "138", "139", "140", "141", "142", "143", "144", "145", "146", "147", "148", "149", "150", "151", "152", "153", "154", "155", "156", "157", "158", "159", "160", "161", "162", "163", "164", "165", "166", "167", "168", "169", "170", "171", "172", "173", "174", "175", "176", "177", "178", "179", "180", "181", "182", "183", "184", "185", "186", "187", "188", "189", "190", "191", "192", "193", "194", "195", "196", "197", "198", "199", "200", "201", "202", "203", "204", "205", "206", "207", "208", "209", "210", "211", "212", "213", "214", "215", "216", "217", "218", "219", "220", "221", "222", "223", "224", "225", "226", "227", "228", "229", "230", "231", "232", "233", "234", "235", "236", "237", "238", "239", "240", "241", "242", "243", "244", "245", "246", "247", "248", "249", "250", "251", "252", "253", "254"]
    }
}
```