@srenatus 2023-06-21
```
goos: darwin
goarch: amd64
pkg: github.com/open-policy-agent/opa/ast
cpu: Intel(R) Core(TM) i9-9880H CPU @ 2.30GHz
                            │    main.bench     │               pr.bench               │
                            │      sec/op       │    sec/op     vs base                │
ObjectFind/5_5-16                  77.12n ± ∞ ¹   75.63n ± ∞ ¹         ~ (p=0.841 n=5)
ObjectFind/5_50-16                 81.61n ± ∞ ¹   83.99n ± ∞ ¹         ~ (p=0.421 n=5)
ObjectFind/5_500-16                82.70n ± ∞ ¹   81.14n ± ∞ ¹         ~ (p=0.690 n=5)
ObjectFind/5_5000-16               80.92n ± ∞ ¹   79.46n ± ∞ ¹         ~ (p=0.548 n=5)
ObjectFind/50_5-16                 86.08n ± ∞ ¹   86.43n ± ∞ ¹         ~ (p=1.000 n=5)
ObjectFind/50_50-16                87.60n ± ∞ ¹   86.00n ± ∞ ¹         ~ (p=1.000 n=5)
ObjectFind/50_500-16               86.36n ± ∞ ¹   84.76n ± ∞ ¹         ~ (p=0.690 n=5)
ObjectFind/50_5000-16              85.69n ± ∞ ¹   91.84n ± ∞ ¹         ~ (p=0.056 n=5)
ObjectFind/500_5-16                78.39n ± ∞ ¹   84.21n ± ∞ ¹    +7.42% (p=0.008 n=5)
ObjectFind/500_50-16               81.87n ± ∞ ¹   89.61n ± ∞ ¹    +9.45% (p=0.008 n=5)
ObjectFind/500_500-16              80.18n ± ∞ ¹   94.39n ± ∞ ¹   +17.72% (p=0.008 n=5)
ObjectFind/500_5000-16             85.23n ± ∞ ¹   89.56n ± ∞ ¹    +5.08% (p=0.016 n=5)
ObjectFind/5000_5-16               80.58n ± ∞ ¹   84.12n ± ∞ ¹         ~ (p=0.151 n=5)
ObjectFind/5000_50-16              80.22n ± ∞ ¹   84.85n ± ∞ ¹    +5.77% (p=0.016 n=5)
ObjectFind/5000_500-16             80.75n ± ∞ ¹   89.31n ± ∞ ¹         ~ (p=0.095 n=5)
ObjectFind/5000_5000-16            84.28n ± ∞ ¹   83.25n ± ∞ ¹         ~ (p=0.841 n=5)
LazyObjectFind/5_5-16            3635.00n ± ∞ ¹   21.48n ± ∞ ¹   -99.41% (p=0.008 n=5)
LazyObjectFind/5_50-16          19832.00n ± ∞ ¹   22.84n ± ∞ ¹   -99.88% (p=0.008 n=5)
LazyObjectFind/5_500-16        188496.00n ± ∞ ¹   23.90n ± ∞ ¹   -99.99% (p=0.008 n=5)
LazyObjectFind/5_5000-16      1842062.00n ± ∞ ¹   27.21n ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/50_5-16           3357.00n ± ∞ ¹   22.94n ± ∞ ¹   -99.32% (p=0.008 n=5)
LazyObjectFind/50_50-16         19506.00n ± ∞ ¹   23.14n ± ∞ ¹   -99.88% (p=0.008 n=5)
LazyObjectFind/50_500-16       184745.00n ± ∞ ¹   23.59n ± ∞ ¹   -99.99% (p=0.008 n=5)
LazyObjectFind/50_5000-16     1684053.00n ± ∞ ¹   25.96n ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/500_5-16          3357.00n ± ∞ ¹   22.42n ± ∞ ¹   -99.33% (p=0.008 n=5)
LazyObjectFind/500_50-16        18929.00n ± ∞ ¹   24.49n ± ∞ ¹   -99.87% (p=0.008 n=5)
LazyObjectFind/500_500-16      158457.00n ± ∞ ¹   24.36n ± ∞ ¹   -99.98% (p=0.008 n=5)
LazyObjectFind/500_5000-16    1704599.00n ± ∞ ¹   26.04n ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5000_5-16         3335.00n ± ∞ ¹   22.15n ± ∞ ¹   -99.34% (p=0.008 n=5)
LazyObjectFind/5000_50-16       18299.00n ± ∞ ¹   22.77n ± ∞ ¹   -99.88% (p=0.008 n=5)
LazyObjectFind/5000_500-16     158947.00n ± ∞ ¹   24.58n ± ∞ ¹   -99.98% (p=0.008 n=5)
LazyObjectFind/5000_5000-16   1591124.00n ± ∞ ¹   25.82n ± ∞ ¹  -100.00% (p=0.008 n=5)
geomean                            2.335µ         45.21n         -98.06%
¹ need >= 6 samples for confidence interval at level 0.95

                            │  main.bench   │                 pr.bench                  │
                            │     B/op      │     B/op       vs base                    │
ObjectFind/5_5-16               16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5_50-16              16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5_500-16             16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5_5000-16            16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/50_5-16              16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/50_50-16             16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/50_500-16            16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/50_5000-16           16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/500_5-16             16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/500_50-16            16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/500_500-16           16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/500_5000-16          16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5000_5-16            16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5000_50-16           16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5000_500-16          16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5000_5000-16         16.00 ± ∞ ¹     16.00 ± ∞ ¹         ~ (p=1.000 n=5) ²
LazyObjectFind/5_5-16         2.307Ki ± ∞ ¹   0.000Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5_50-16        13.48Ki ± ∞ ¹    0.00Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5_500-16       149.7Ki ± ∞ ¹     0.0Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5_5000-16      1.624Mi ± ∞ ¹   0.000Mi ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/50_5-16        2.307Ki ± ∞ ¹   0.000Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/50_50-16       13.48Ki ± ∞ ¹    0.00Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/50_500-16      149.4Ki ± ∞ ¹     0.0Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/50_5000-16     1.620Mi ± ∞ ¹   0.000Mi ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/500_5-16       2.307Ki ± ∞ ¹   0.000Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/500_50-16      13.47Ki ± ∞ ¹    0.00Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/500_500-16     148.7Ki ± ∞ ¹     0.0Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/500_5000-16    1.611Mi ± ∞ ¹   0.000Mi ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5000_5-16      2.306Ki ± ∞ ¹   0.000Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5000_50-16     13.46Ki ± ∞ ¹    0.00Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5000_500-16    148.4Ki ± ∞ ¹     0.0Ki ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5000_5000-16   1.609Mi ± ∞ ¹   0.000Mi ± ∞ ¹  -100.00% (p=0.008 n=5)
geomean                         928.5                        ?                      ³ ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean
⁴ ratios must be >0 to compute geomean

                            │  main.bench  │                 pr.bench                 │
                            │  allocs/op   │  allocs/op    vs base                    │
ObjectFind/5_5-16              1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5_50-16             1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5_500-16            1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5_5000-16           1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/50_5-16             1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/50_50-16            1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/50_500-16           1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/50_5000-16          1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/500_5-16            1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/500_50-16           1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/500_500-16          1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/500_5000-16         1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5000_5-16           1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5000_50-16          1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5000_500-16         1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
ObjectFind/5000_5000-16        1.000 ± ∞ ¹    1.000 ± ∞ ¹         ~ (p=1.000 n=5) ²
LazyObjectFind/5_5-16          37.00 ± ∞ ¹     0.00 ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5_50-16         260.0 ± ∞ ¹      0.0 ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5_500-16       2.517k ± ∞ ¹   0.000k ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5_5000-16      25.03k ± ∞ ¹    0.00k ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/50_5-16         37.00 ± ∞ ¹     0.00 ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/50_50-16        260.0 ± ∞ ¹      0.0 ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/50_500-16      2.517k ± ∞ ¹   0.000k ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/50_5000-16     25.03k ± ∞ ¹    0.00k ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/500_5-16        37.00 ± ∞ ¹     0.00 ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/500_50-16       260.0 ± ∞ ¹      0.0 ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/500_500-16     2.517k ± ∞ ¹   0.000k ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/500_5000-16    25.03k ± ∞ ¹    0.00k ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5000_5-16       37.00 ± ∞ ¹     0.00 ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5000_50-16      260.0 ± ∞ ¹      0.0 ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5000_500-16    2.517k ± ∞ ¹   0.000k ± ∞ ¹  -100.00% (p=0.008 n=5)
LazyObjectFind/5000_5000-16   25.03k ± ∞ ¹    0.00k ± ∞ ¹  -100.00% (p=0.008 n=5)
geomean                        29.70                       ?                      ³ ⁴
¹ need >= 6 samples for confidence interval at level 0.95
² all samples are equal
³ summaries must be >0 to compute geomean
⁴ ratios must be >0 to compute geomean
```

LGTM! Thanks for adding benchmarks!